moved transform.c file to src and changed the corresponding relative path in all main.cpp files
