+++ src/top_model/main.cpp
@@ -23,7 +23,7 @@
 #include "../../include/atomics/receivercadmium.hpp"
 #include "../../include/atomics/subnetcadmium.hpp"
 
-#include "../../transform.cpp"
+#include "../transform.cpp"
 
 
 using namespace std;
