+++ test/src/sender/main.cpp
@@ -24,7 +24,7 @@
 #include "../../../lib/iestream.hpp"
 #include "../../../include/data_structures/message.hpp"
 #include "../../../include/atomics/sendercadmium.hpp"
-#include "../../../transform.cpp"
+#include "../../../src/transform.cpp"
 
 using namespace std;
 using hclock = chrono::high_resolution_clock;
