+++ test/src/subnet/main.cpp
@@ -17,7 +17,7 @@
 #include "../../../lib/iestream.hpp"
 #include "../../../include/data_structures/message.hpp"
 #include "../../../include/atomics/subnetcadmium.hpp"
-#include "../../../transform.cpp"
+#include "../../../src/transform.cpp"
 
 using namespace std;
 using hclock = chrono::high_resolution_clock;
